Update safe_withdraw.js
withdrawal-memo in buildSafeTransaction is not the actual memo. It only exists in Mixin mainnet.

e.g. For a EOS transaction. If you specify the memo in the field of 'withdrawal-memo' (which is an argument of buildSafeTransaction), you won't receive your fund on the other side. It turns out you can only set the memo in 'recipients' to initiate a successful withdrawal. 

Failed example (No Memo):
https://mixin.space/tx/c4c145e3d8c562b13002cc080cfdd5b89a6978a877fcc678ddbd3f7d978965f5
https://eosflare.io/tx/FE73D7CCC4319C4EA931657CA402C57886F4AF87203AAA9B7E8FE183ECD8EEF2

Success example (With Memo): 
https://mixin.space/tx/4430151efdb35cc26b6863e7048417beb3ad88e3b45149efdb55d73f576bdd26
https://eosflare.io/tx/8CDB5B8DBB97718FDB5B71570B9316AACCB1E0D59BD2841EDCE8EC66057A90A3